services/backend/requirements.txt
ansicon==1.89.0
apiview==1.3.25
asgiref==3.6.0
async-timeout==4.0.2
beautifulsoup4==4.12.0
blessed==1.20.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
charset-normalizer==3.1.0
cloudipsp==1.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.7
django-cors-headers==3.14.0
django-environ==0.10.0
django-filter==23.1
django-graylog==0.7.0
django-import-export==3.1.0
django-modeltranslation==0.18.9
django-rest-multiple-models==2.1.3
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg==1.21.5
et-xmlfile==1.1.0
facebook-sdk==3.1.0
fuzzywuzzy==0.18.0
gevent==22.10.2
google-auth==2.16.3
graypy==2.1.0
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
inquirer==3.1.3
itypes==1.2.0
Jinja2==3.1.2
jinxed==1.2.0
Levenshtein==0.20.9
MarkupPy==1.14
MarkupSafe==2.1.2
msgpack==1.0.5
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.0
psycopg2==2.9.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
python-dotenv==1.0.0
python-editor==1.0.4
python-Levenshtein==0.20.9
python3-openid==3.2.0
pytz==2023.2
pyuseragents==1.0.5
PyYAML==6.0
rapidfuzz==2.13.7
readchar==4.0.5
redis==4.5.3
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
safeIO==1.2
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.4.0
soupsieve==2.4
sqlparse==0.4.3
tablib==3.4.0
translatepy==2.3
typing_extensions==4.5.0
tzdata==2023.2
uritemplate==4.1.1
urllib3==1.26.15
wcwidth==0.2.6
xlrd==2.0.1
xlwt==1.3.0
zope.event==4.6
zope.interface==6.0
